restructured eval programs
